# src/main/resources/static/swagger.yml
openapi: 3.0.3
info:
  title: Modam API
  description: 커플 공동 금융 서비스 전체 API 명세
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Server
...
paths:
...
  /api/accounts/{accountId}:
    get:
      tags: [Account]
      summary: 계좌 단건 조회
      operationId: getAccount
      parameters:
        - $ref: '#/components/parameters/accountIdPath'
      responses:
        '200':
          description: 계좌 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GlobalResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountResponse'

    patch:
      tags: [Account]
      summary: 계좌 정보 수정
      operationId: updateAccount
      parameters:
        - $ref: '#/components/parameters/accountIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequest'
      responses:
        '200':
          description: 계좌 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GlobalResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountResponse'

    delete:
      tags: [Account]
      summary: 계좌 해지 (status → CLOSED)
      operationId: closeAccount
      parameters:
        - $ref: '#/components/parameters/accountIdPath'
      responses:
        '200':
          description: 계좌 해지 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GlobalResponse'
                  - type: object
                    properties:
                      data:
                        type: string
                        example: "계좌가 해지되었습니다."
...
  /api/invites/accounts/{accountId}:
    post:
      tags: [Invite]
      summary: 모임 통장 파트너 초대 (최대 2명)
      operationId: inviteMember
      parameters:
        - $ref: '#/components/parameters/accountIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteRequest'
      responses:
        '200':
          description: 초대 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GlobalResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InviteResponse'
        '400':
          description: 잘못된 요청 (개인 계좌, 최대 인원 초과, 중복 초대)
  /api/invites/{accountMemberId}/accept:
    patch:
      tags: [Invite]
      summary: 모임 통장 초대 수락
      operationId: acceptInvite
      parameters:
        - name: accountMemberId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 초대 수락 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GlobalResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InviteResponse'
        '400':
          description: 존재하지 않는 초대 정보

  # ── Card ────────────────────────────────────────────────────
...
  /api/spend-records/transaction/{transactionId}:
    get:
      tags: [SpendRecord]
      summary: 거래 내역의 소비 기록 조회
      operationId: getSpendRecordByTransaction
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 소비 기록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GlobalResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SpendRecordResponse'
...
  /api/transactions:
    post:
      tags: [Transaction]
      summary: 거래 내역 생성 (입금/출금/결제)
      description: |
        JWT 인증 필요. 인증 토큰에서 회원 정보를 추출하므로 요청 본문에 memberId 불필요.
        - PAYMENT 타입인 경우 cardId 필수
        - 출금/결제 시 잔액 부족이면 400 반환
      operationId: createTransaction
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '200':
          description: 거래 처리 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GlobalResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: 잘못된 요청 (잔액 부족, PAYMENT인데 cardId 없음, 카드 소유권 불일치 등)
        '401':
          description: 인증 토큰 없음 또는 만료
        '403':
          description: 해당 계좌에 대한 접근 권한 없음
  /api/transactions/{accountId}:
    get:
      tags: [Transaction]
      summary: 거래 내역 조회 (무한 스크롤 – No-offset)
      description: ID 내림차순으로 정렬. lastTransactionId를 커서로 사용하는 커서 기반 페이지네이션.
      operationId: getTransactions
      parameters:
        - $ref: '#/components/parameters/accountIdPath'
        - name: lastTransactionId
          in: query
          required: false
          description: 마지막으로 본 거래 ID (첫 요청 시 생략)
          schema:
            type: integer
            format: int64
        - name: size
          in: query
          required: false
          description: 한 번에 가져올 건수 (기본 10)
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 거래 내역 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GlobalResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: 존재하지 않는 계좌 ID

# ─────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: "세션 기반 인증 (Spring Security formLogin, JSESSIONID 쿠키)"
  parameters:
...
    accountIdPath:
      name: accountId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    GlobalResponse:
      type: object
      description: 모든 API 공통 응답 래퍼
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          nullable: true
          example: "200"
        message:
          type: string
          example: "success"
        data:
          nullable: true
      required:
        - success
        - message

    # ── Member ─────────────────────────────────────────────────
...
    AccountUpdateRequest:
      type: object
      properties:
        deliveryAddress:
          type: string
          nullable: true
        jobInfo:
          type: string
          nullable: true
        tradePurpose:
          type: string
          nullable: true
        fundSource:
          type: string
          nullable: true
        spendLimitAmount:
          type: integer
          format: int64
          nullable: true
          example: 500000
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        accountNumber:
          type: string
          example: "A1B2C3D4E5F6G7H8"
        accountType:
          type: string
          example: "GROUP"
          enum: [PERSONAL, GROUP]
        status:
          type: string
          example: "ACTIVE"
          enum: [ACTIVE, FREEZE, CLOSED]
        balance:
          type: integer
          format: int64
          example: 150000
        availableBalance:
          type: integer
          format: int64
          example: 150000
        spendLimitAmount:
          type: integer
          format: int64
          nullable: true
          example: 500000
        createdAt:
          type: string
          format: date-time
...
    InviteRequest:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
          example: 2
          description: 초대할 회원 ID
      required:
        - memberId
    InviteResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        memberId:
          type: integer
          format: int64
          example: 2
        memberName:
          type: string
          example: "홍길동"
        inviteStatus:
          type: string
          example: "WAIT"
          enum: [WAIT, ACCEPT, REJECT]
        totalDeposit:
          type: integer
          format: int64
          example: 0

    # ── Card ───────────────────────────────────────────────────
...
    SpendRecordResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        transactionId:
          type: integer
          format: int64
          example: 10
        imageUrl:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
          example: "치킨 시켜먹음"
        emoticon:
          type: string
          nullable: true
          example: "🍗"
        createdAt:
          type: string
          format: date-time

    # ── Transaction ────────────────────────────────────────────
    TransactionCreateRequest:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
          example: 1
        memberId:
          type: integer
          format: int64
          example: 2
        cardId:
          type: integer
          format: int64
          nullable: true
          description: "카드 결제 시에만 포함 (txType=PAYMENT 이면 필수)"
        txType:
          type: string
          example: "DEPOSIT"
          enum: [DEPOSIT, WITHDRAW, PAYMENT]
        amount:
          type: integer
          format: int64
          example: 30000
          minimum: 1
          description: 거래 금액 (0보다 커야 함)
        merchantName:
          type: string
          nullable: true
          example: "BBQ 강남점"
        category:
          type: string
          nullable: true
          example: "식비"
      required:
        - accountId
        - memberId
        - txType
        - amount
    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
          example: 1
        txType:
          type: string
          example: "PAYMENT"
          enum: [DEPOSIT, WITHDRAW, PAYMENT]
        amount:
          type: integer
          format: int64
          example: 30000
        afterBalance:
          type: integer
          format: int64
          example: 120000
        merchantName:
          type: string
          nullable: true
          example: "BBQ 강남점"
        category:
          type: string
          nullable: true
          example: "식비"
        createdAt:
          type: string
          format: date-time